fix: fix comment block syntax & folding
This corrects the issue with #comment-start/#comment-end/#cs/#ce failing to end syntax highlighting when mixed together, and adds proper code folding functionality when comments are added.

I was able to get the syntax highlighting for the blocks to work together by combining their Regex in the .tmLanguage file. To fix the folding I simply added them to folding.markers in the language configuration file.

Addresses issue #89.